+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Lookup results are now slightly more intelligent:
   - if empty, show first level results
   - otherwise show first all that start with query sorted, then those that contain it sorted
+  - if query also equal to current note, also show siblings
 - Lookup query matches are now highlighted
 
 ## 1.3.0